requirements.txt
aws-cdk-lib==2.149.0
constructs>=10.0.0,<11.0.0
pydantic==2.8.2
pydantic-settings==2.3.4
